refactor: create layers and categories feature directories (Phase 5d)
Move layers and categories components to new feature-based directories:

Layers feature (src/features/layers/):
- components/LayerPanel.tsx, ActiveLayerPanel.tsx
- hooks/useAdvancedLayerMode.ts
- Barrel exports

Categories feature (src/features/categories/):
- components/CategoriesPanel.tsx
- Barrel exports

Re-export stubs at old locations:
- src/components/Sidebar/{LayerPanel,ActiveLayerPanel,CategoriesPanel}.tsx
- src/hooks/useAdvancedLayerMode.ts

These re-export stubs will be removed in Phase 6.

Part of folder structure reorganization Phase 5.
