+++ src/features/categories/components/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for categories components
+export { CategoriesPanel } from './CategoriesPanel';
