+++ src/features/categories/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for categories feature
+export * from './components';
