+++ src/features/layers/components/index.ts
@@ -1,0 +1,3 @@
+// Barrel export for layers components
+export { LayerPanel } from './LayerPanel';
+export { ActiveLayerPanel } from './ActiveLayerPanel';
