+++ src/features/layers/hooks/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for layers hooks
+export { useAdvancedLayerMode } from './useAdvancedLayerMode';
