+++ src/features/layers/index.ts
@@ -1,0 +1,3 @@
+// Barrel export for layers feature
+export * from './components';
+export * from './hooks';
