+++ src/hooks/useAdvancedLayerMode.ts
@@ -1,10 +1,3 @@
-import { useLayoutStore } from '../core/store';
-
-/**
- * Returns true if advanced layer mode should be shown.
- * True when user has 2+ layers.
- */
-export function useAdvancedLayerMode(): boolean {
-  const layerCount = useLayoutStore((state) => state.layout.layers.length);
-  return layerCount > 1;
-}
+// Re-export from new location for backward compatibility
+// TODO: Remove in Phase 6 - update all imports to use features/layers
+export { useAdvancedLayerMode } from '../features/layers';
